# util/deephash: optimize hashing of byte arrays, reduce allocs in Hash