@crawshaw 2021-07-06
What about my comment about isolating the unsafe to its own package?

Does the `reflect.Value.Bytes` suggestion deserve an upstream issue?