@bradfitz 2021-07-06
> What about my comment about isolating the unsafe to its own package?

Doesn't seem worth it. This is already a pretty low-level reflect-y package. I can't think of a good API boundary between the two.

Plus maybe we can find a safe way to to do it later.

> Does the reflect.Value.Bytes suggestion deserve an upstream issue?

Sure: https://github.com/golang/go/issues/47066