@crawshaw 2021-07-06
Can you help me understand why using unsafe here is worth it? The micro benchmark doesn’t explain to me the value.