@josharian 2021-07-06
What type is this for? Can we make it implement AppendTo instead?

I went through this fight a bunch of different ways and eventually gave up, b/c I didn't want to use unsafe.