@bradfitz 2021-07-07
> Can you help me understand why using unsafe here is worth it? The micro benchmark doesn’t explain to me the value.

The context is replacing control's JSON MapResponse minifier current JSON checksumming mechanism with deephash. So deephash would be seeing many more types it's never been optimized for before, including everything under `tailcfg.MapResponse`, notably lots of `tailcfg.Node` (which include a few array types that are currently unnecessarily costly)