@bradfitz 2021-07-07
I forgot to cross-reference https://github.com/tailscale/corp/issues/2130