@bradfitz 2021-07-07
> What type is this for? Can we make it implement AppendTo instead?

I'm not a huge fan of the AppendTo hack.

> I went through this fight a bunch of different ways and eventually gave up, b/c I didn't want to use unsafe.

I'll try a bit more to do this without unsafe, even if it's not quite as fast. But it's easy to do better than what we do now.