@bradfitz 2021-07-07
I found a way to do this without unsafe, at least for all arrays <= the scratch space size, using reflect.Copy.

Updating this PR after some other cleanups & fixes.