@bradfitz 2021-07-07
PTAL; now without unsafe.